1branch 1environment. A whole new world
Moved all of thow's code to a single environment, "base"

Issues summed up:

1
We were reverting to the wrong commit to find the Thow code.
"made new code for usb connection" is the wrong code
"resolving pymeasure install issue" has the actuall thow code (and presumably some Xiaokun updates)

2
We didn't have the atto_device package installed.
It's unnesecary when we have the bypass, but the other classes within the heterodyne procedure file (that are unused) need that class.

3
we were unsure of the correct virtual env. Thow used the .venv, but we wanted to simplify and delete that

4
it was saying com port 4 not found. This was because the usb cable was unplugged.